requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

attrs==22.2.0
cloudinary==1.33.0
cryptography==41.0.3
Django==4.1.5
django-filter==23.1
docutils==0.14
ipython==8.8.0
ipywidgets==8.0.4
Jinja2==3.1.2
numpy==1.24.1
packaging==21.3
Pillow==9.5.0
pip==23.2.1
pyOpenSSL==23.2.0
pyparsing==3.0.9
python-dotenv==1.0.0
pywin32==305
setuptools==65.5.0
tornado==6.2
wheel==0.38.4
